+++ src/sparkx/Particle.py
@@ -19,7 +19,7 @@ class Particle:
 
     The member variables of the Particle class are the quantities in the
     Oscar2013/Oscar2013Extended/Oscar2013Extended_IC/Oscar2013Extended_Photons/
-    ASCIICustom or JETSCAPE hadron/parton output. If they are not set,
+    ASCII or JETSCAPE hadron/parton output. If they are not set,
     they stay :code:`np.nan` to throw an error if one tries to access a non existing
     quantity.
     If a particle with an unknown PDG is provided, a warning is thrown and and
@@ -212,7 +212,7 @@ class Particle:
 
     * "Oscar2013Extended_Photons"
 
-    * "ASCIICustom"
+    * "ASCII"
 
     * "JETSCAPE"
 
@@ -241,7 +241,7 @@ class Particle:
         self,
         input_format: Optional[str] = None,
         particle_array: Optional[np.ndarray] = None,
-        attribute_list: List[str] = []
+        attribute_list: List[str] = [],
     ) -> None:
         self.data_: np.ndarray = np.array(25 * [np.nan], dtype=float)
         self.pdg_valid: bool = False
@@ -250,18 +250,25 @@ class Particle:
             (input_format is None) and (particle_array is not None)
         ):
             raise ValueError("'input_format' or 'particle_array' not given")
-        
-        if attribute_list is None and input_format == "ASCIICustom":
+
+        if attribute_list is None and input_format == "ASCII":
             raise ValueError("'attribute_list' not given")
 
-        if attribute_list!=[] and input_format != "ASCIICustom":
-            raise ValueError("'OscarCustom' format requires 'attribute_list' to be given.")
+        if attribute_list != [] and input_format != "ASCII":
+            raise ValueError(
+                "'OscarCustom' format requires 'attribute_list' to be given."
+            )
 
         if (input_format is not None) and (particle_array is not None):
-            self.__initialize_from_array(input_format, particle_array, attribute_list)
+            self.__initialize_from_array(
+                input_format, particle_array, attribute_list
+            )
 
     def __initialize_from_array(
-        self, input_format: str, particle_array: np.ndarray, attribute_list: List[str]
+        self,
+        input_format: str,
+        particle_array: np.ndarray,
+        attribute_list: List[str],
     ) -> None:
         """
         Initialize instance attributes based on the provided input format and array, and optionally an atttribute list.
@@ -274,7 +281,7 @@ class Particle:
             - "Oscar2013Extended"
             - "Oscar2013Extended_IC"
             - "Oscar2013Extended_Photons"
-            - "ASCIICustom"
+            - "ASCII"
             - "JETSCAPE"
 
         particle_array : numpy.ndarray
@@ -304,29 +311,29 @@ class Particle:
         # second entry: index in line
         attribute_mapping = {
             "Allfields": {
-                "t": [0,0],
-                "x": [1,0],
-                "y": [2,0],
-                "z": [3,0],
-                "mass": [4,0],
-                "E": [5,0],
-                "px": [6,0],
-                "py": [7,0],
-                "pz_": [8,0],
-                "pdg": [9,0],
-                "ID": [11,0],
-                "charge": [12,0],
-                "ncoll": [13,0],
-                "form_time": [14,0],
-                "xsecfac": [15,0],
-                "proc_id_origin": [16,0],
-                "proc_type_origin": [17,0],
-                "t_last_coll": [18,0],
-                "pdg_mother1": [19,0],
-                "pdg_mother2": [20,0],
-                "status_": [21,0],
-                "baryon_number": [22,0],
-                "strangeness": [23,0],
+                "t": [0, 0],
+                "x": [1, 0],
+                "y": [2, 0],
+                "z": [3, 0],
+                "mass": [4, 0],
+                "E": [5, 0],
+                "px": [6, 0],
+                "py": [7, 0],
+                "pz_": [8, 0],
+                "pdg": [9, 0],
+                "ID": [11, 0],
+                "charge": [12, 0],
+                "ncoll": [13, 0],
+                "form_time": [14, 0],
+                "xsecfac": [15, 0],
+                "proc_id_origin": [16, 0],
+                "proc_type_origin": [17, 0],
+                "t_last_coll": [18, 0],
+                "pdg_mother1": [19, 0],
+                "pdg_mother2": [20, 0],
+                "status_": [21, 0],
+                "baryon_number": [22, 0],
+                "strangeness": [23, 0],
             },
             "Oscar2013": {
                 "t_": [0, 0],
@@ -423,19 +430,31 @@ class Particle:
                 "pz_": [8, 6],
             },
         }
-        if input_format == "ASCIICustom":
+        if input_format == "ASCII":
             mapping_dict = {}
             for attr in attribute_list:
-                mapping_dict[attr] = [attribute_mapping["Allfields"][attr][0], list(attribute_list).index(attr)]
-            attribute_mapping["ASCIICustom"] = mapping_dict
-        if input_format in attribute_mapping or input_format=="ASCIICustom":
-            if (input_format == "ASCIICustom" or len(particle_array) == len(attribute_mapping[input_format]) or (input_format in ["Oscar2013Extended","Oscar2013Extended_IC"]\
-                    and len(particle_array) <=  len(attribute_mapping[input_format])\
-                    and len(particle_array) >=  len(attribute_mapping[input_format])-2)):
+                mapping_dict[attr] = [
+                    attribute_mapping["Allfields"][attr][0],
+                    list(attribute_list).index(attr),
+                ]
+            attribute_mapping["ASCII"] = mapping_dict
+        if input_format in attribute_mapping or input_format == "ASCII":
+            if (
+                input_format == "ASCII"
+                or len(particle_array) == len(attribute_mapping[input_format])
+                or (
+                    input_format
+                    in ["Oscar2013Extended", "Oscar2013Extended_IC"]
+                    and len(particle_array)
+                    <= len(attribute_mapping[input_format])
+                    and len(particle_array)
+                    >= len(attribute_mapping[input_format]) - 2
+                )
+            ):
                 for attribute, index in attribute_mapping[input_format].items():
                     if len(particle_array) <= (index[1]):
                         continue
-                    if input_format == "ASCIICustom":
+                    if input_format == "ASCII":
                         attribute = attribute + "_"
                     # Type casting for specific attributes. Although everything is saved as a float, we will only read in int data for int fields
                     # to ensure similar behaving as if we were reading in data
@@ -500,9 +519,9 @@ class Particle:
         if not self.pdg_valid:
             if np.isnan(self.pdg):
                 warnings.warn(
-                "No PDG code given! "
-                + "All properties extracted from the PDG are set to default values."
-            )
+                    "No PDG code given! "
+                    + "All properties extracted from the PDG are set to default values."
+                )
             else:
                 warnings.warn(
                     "The PDG code "