+++ src/sparkx/loader/JetscapeLoader.py
@@ -157,7 +157,7 @@ class JetscapeLoader(BaseLoader):
             self.set_particle_list(kwargs),
             self.num_events_,
             self.num_output_per_event_,
-            []
+            [],
         )
 
     # PRIVATE CLASS METHODS
@@ -423,7 +423,7 @@ class JetscapeLoader(BaseLoader):
         particle_list: List[List[Particle]] = []
         data: List[Particle] = []
         num_read_lines = self.__get_num_read_lines()
-        cut_events=0
+        cut_events = 0
         with open(self.PATH_JETSCAPE_, "r") as jetscape_file:
             self._skip_lines(jetscape_file)
 
@@ -432,23 +432,34 @@ class JetscapeLoader(BaseLoader):
                 if not line:
                     raise IndexError("Index out of range of JETSCAPE file")
                 elif "#" in line and "sigmaGen" in line:
-                    old_data_len=len(data)
+                    old_data_len = len(data)
                     if "filters" in self.optional_arguments_.keys():
                         data = self.__apply_kwargs_filters(
                             [data], kwargs["filters"]
                         )[0]
-                        if (len(data) != 0 or old_data_len==0):
+                        if len(data) != 0 or old_data_len == 0:
                             self.num_output_per_event_[len(particle_list)] = (
-                                len(particle_list)+1,
+                                len(particle_list) + 1,
                                 len(data),
                             )
                         else:
-                            self.num_output_per_event_ = np.atleast_2d(np.delete(self.num_output_per_event_, len(particle_list), axis=0))                  
+                            self.num_output_per_event_ = np.atleast_2d(
+                                np.delete(
+                                    self.num_output_per_event_,
+                                    len(particle_list),
+                                    axis=0,
+                                )
+                            )
                             if self.num_output_per_event_.shape[0] == 0:
                                 self.num_output_per_event_ = np.array([])
-                            elif len(particle_list) < self.num_output_per_event_.shape[0]:
-                                self.num_output_per_event_[len(particle_list):, 0] -= 1
-                    if (len(data) != 0 or old_data_len==0 ):
+                            elif (
+                                len(particle_list)
+                                < self.num_output_per_event_.shape[0]
+                            ):
+                                self.num_output_per_event_[
+                                    len(particle_list) :, 0
+                                ] -= 1
+                    if len(data) != 0 or old_data_len == 0:
                         particle_list.append(data)
                     else:
                         cut_events = cut_events + 1
@@ -477,23 +488,36 @@ class JetscapeLoader(BaseLoader):
                     if int(line_list[2]) == first_event_header:
                         continue
                     else:
-                        old_data_len=len(data)
+                        old_data_len = len(data)
                         if "filters" in self.optional_arguments_.keys():
                             data = self.__apply_kwargs_filters(
                                 [data], kwargs["filters"]
                             )[0]
-                            if (len(data) != 0 or old_data_len==0):
-                                self.num_output_per_event_[len(particle_list)] = (
-                                    len(particle_list)+1,
+                            if len(data) != 0 or old_data_len == 0:
+                                self.num_output_per_event_[
+                                    len(particle_list)
+                                ] = (
+                                    len(particle_list) + 1,
                                     len(data),
                                 )
                             else:
-                                self.num_output_per_event_ = np.atleast_2d(np.delete(self.num_output_per_event_, len(particle_list), axis=0))                  
+                                self.num_output_per_event_ = np.atleast_2d(
+                                    np.delete(
+                                        self.num_output_per_event_,
+                                        len(particle_list),
+                                        axis=0,
+                                    )
+                                )
                                 if self.num_output_per_event_.shape[0] == 0:
                                     self.num_output_per_event_ = np.array([])
-                                elif len(particle_list) < self.num_output_per_event_.shape[0]:
-                                    self.num_output_per_event_[len(particle_list):, 0] -= 1
-                        if (len(data) != 0 or old_data_len==0 ):
+                                elif (
+                                    len(particle_list)
+                                    < self.num_output_per_event_.shape[0]
+                                ):
+                                    self.num_output_per_event_[
+                                        len(particle_list) :, 0
+                                    ] -= 1
+                        if len(data) != 0 or old_data_len == 0:
                             particle_list.append(data)
                         else:
                             cut_events = cut_events + 1
@@ -504,8 +528,8 @@ class JetscapeLoader(BaseLoader):
                     )
                     particle = Particle("JETSCAPE", np.asarray(line_list))
                     data.append(particle)
-        
-        self.num_events_=self.num_events_-cut_events
+
+        self.num_events_ = self.num_events_ - cut_events
         # Correct num_output_per_event and num_events
         if not kwargs or "events" not in self.optional_arguments_.keys():
             if len(particle_list) != self.num_events_: